Added Windows notepad tool use.

+++ src/api.py
@@ -32,6 +32,8 @@
 # POSSIBILITY OF SUCH DAMAGE.
 
 import os
+import os.path
+import platform
 import random
 import subprocess
 from sys import exit
@@ -196,19 +198,39 @@ class API:
                 else:
                     raise EditorNotFound()
             except EditorNotFound:
-                print("Text editor not found; set the 'EDITOR'" +
+                if platform.system() == "Windows":
+                    editor = "notepad"
+                    windowsfilename = f"{self.target.reference}"
+                    windowsfilename = windowsfilename.replace(":","-")
+                else:
+                    print("Text editor not found; set the 'EDITOR'" +
                       "environmental variable and try again")
-                exit()
+                    exit()
 
-        self.filename = self.cache_path / f"{self.target.reference}"
-        subprocess.run([editor, self.filename])
+        if platform.system() == "Windows":
+            self.filename = self.cache_path / windowsfilename
+        else:
+            self.filename = self.cache_path / f"{self.target.reference}"
+
+        if os.path.isfile(self.filename) == False:
+            with open(self.filename, "w") as file:
+                subprocess.run([editor, self.filename])
+        else:
+            subprocess.run([editor, self.filename])
+
+        withextension = f"{self.filename}" + ".txt"
 
         if not self.filename.exists():
             text = ""
         else:
-            with open(self.filename, "r") as file:
-                text = file.readlines()
-                text = "".join(text)
+            if platform.system() == "Windows":
+                with open(withextension, "r") as file:
+                    text = file.readlines()
+                    text = "".join(text)
+            else:
+                with open(self.filename, "r") as file:
+                    text = file.readlines()
+                    text = "".join(text)
 
             # Vim Automatically Adds a Newline at the End of the File when
             # you save it. (Unless you set :nofixeol and set :nofixendofline -
@@ -218,7 +240,13 @@ class API:
             if (editor in ("vim", "nvim", "nano")) and text[-1] == "\n":
                 text = text[:-1]
 
-            os.remove(self.filename)
+            if os.path.isfile(self.filename):
+                os.remove(self.filename)
+
+            # withextension = f"{self.filename}" + ".txt"
+            if os.path.isfile(withextension):
+                os.remove(withextension)
+
 
         attempt = Attempt.create(
             random_mode=self.mode,
